requirements.txt
torch
torchvision
numpy
matplotlib
dvc[gs] 
flask
torch
torchvision
pillow
flask-cors
